feat: P2/P3/P4 — freeze compat story, operational hardening, integrator UX (#247)
* feat: P2/P3/P4 — freeze compat story, operational hardening, integrator UX

P2: Freeze Supported Compatibility Story
- Add enable_v1_compat_writes config flag (default true for backward compat)
- Gate best-effort V1 writes in api worker and stream worker behind the flag
- Thread config through flush_stream_batch and run_hyperliquid_ws_stream
- Mark CLI normalize as legacy compatibility in help text
- Document flag in spectraplex.toml.example

P3: Minimal Operational Hardening
- Add adapters/tests/operational_hardening_test.rs with 4 integration tests:
  - materialize_idempotency: re-running materialization on same ingestion run
    does not duplicate Gold records
  - tenant_isolation_cross_owner_target_access: cross-tenant target listing
    and access is properly scoped
  - export_gold_dataset_lifecycle: enqueue, claim, and complete an export job
    for wallet_ledger (Gold)
  - export_job_claim_and_update: worker claim + status update round-trip
- Fix pre-existing bug: STALE_JOB_THRESHOLD_MINUTES i64 → i32 to match
  PostgreSQL make_interval(mins => integer) signature

P4: Minimal Integrator UX
- README: add 'Supported Path vs Legacy Compatibility' section
- README: add Callback HMAC Verification example with Python snippet
- README: add SPECTRAPLEX_ENABLE_V1_COMPAT_WRITES env var note
- design doc: update workstream status table (F, G, H marked Done for MVP)

* fix: address codex review — make idempotency test actually assert on non-zero rows

- Replace Solana tx helper with EVM tx helper that produces ERC20 transfer
  metadata, ensuring materialize_silver_datasets emits token_transfers rows
- Assert count1 > 0 before checking idempotency, so the test would fail if
  duplicate upserts were broken rather than trivially passing with zero rows

+++ adapters/src/v2_repo.rs
@@ -1597,7 +1597,7 @@ impl Repository {
 
     /// Threshold after which a claimed/running job with no heartbeat is
     /// considered abandoned by a dead worker and eligible for reclaim.
-    const STALE_JOB_THRESHOLD_MINUTES: i64 = 5;
+    const STALE_JOB_THRESHOLD_MINUTES: i32 = 5;
 
     // -----------------------------------------------------------------------
     // Networks
